+++ lib/regexp_parser/expression/base.rb
@@ -26,24 +26,17 @@ module Regexp::Expression
 
     # Deprecated. Prefer `#repetitions` which has a more uniform interface.
     def quantity
-      return [nil,nil] unless quantified?
+      return [nil, nil] unless quantified?
+
       [quantifier.min, quantifier.max]
     end
 
     def repetitions
-      @repetitions ||=
-        if quantified?
-          min = quantifier.min
-          max = quantifier.max < 0 ? Float::INFINITY : quantifier.max
-          range = min..max
-          # fix Range#minmax on old Rubies - https://bugs.ruby-lang.org/issues/15807
-          if RUBY_VERSION.to_f < 2.7
-            range.define_singleton_method(:minmax) { [min, max] }
-          end
-          range
-        else
-          1..1
-        end
+      return 1..1 unless quantified?
+
+      min = quantifier.min
+      max = quantifier.max < 0 ? Float::INFINITY : quantifier.max
+      min..max
     end
 
     def greedy?
@@ -74,5 +67,48 @@ module Regexp::Expression
       }
     end
     alias :attributes :to_h
+
+    # Recalculates the nesting_level for this node and children, which reflects
+    # how deep a node is in the tree. Do this at the end of parsing to account
+    # for tree rewrites.
+    # This must be safe against overflows and performant - check the benchmarks.
+    def recursively_update_nesting_levels(base = nesting_level)
+      queue = [base, self]
+
+      until queue.empty?
+        exp = queue.pop
+        exp_lvl = queue.pop
+        exp.nesting_level = exp_lvl
+        exp.quantifier.nesting_level = exp_lvl if exp.quantifier
+        exp.terminal? || exp.expressions.each do |subexp|
+          queue.push(exp_lvl + 1, subexp)
+        end
+      end
+    end
+
+    # Assigns referenced expressions to referring expressions, e.g. if there is
+    # an instance of Backreference::Number, its #referenced_expression is set to
+    # the instance of Group::Capture that it refers to via its number.
+    def assign_referenced_expressions
+      # find all referenceable and referring expressions
+      targets = {}
+      targets[0] = [self] if instance_of?(Root)
+      referrers = []
+      each_expression do |exp|
+        if exp.referential?
+          referrers << exp
+        elsif exp.is_a?(Group::Capture)
+          (targets[exp.identifier] ||= []) << exp
+        end
+      end
+      # assign referenced expressions to referring expressions
+      # (in a second iteration because there might be forward references)
+      referrers.each do |exp|
+        exp.referenced_expressions = targets[exp.reference] || raise(
+          Regexp::Parser::ParserError,
+          "Invalid reference #{exp.reference} at pos #{exp.ts}"
+        )
+      end
+    end
   end
 end