+++ spec/parser/all_spec.rb
@@ -31,4 +31,15 @@ RSpec.describe(Regexp::Parser) do
   specify('parse sequence no quantifier target raises error') do
     expect { RP.parse('abc|?def') }.to raise_error(Regexp::Parser::Error)
   end
+
+  specify('is resistant against stack overflows') do
+    pattern = "#{'(' * 9000}abc#{')' * 9000}" # Ruby's max regexp depth is 4096
+    expect { @root = RP.parse(pattern) }.not_to raise_error
+    expect(@root).to eq @root
+    expect(@root.to_s).to eq(pattern)
+    expect(@root.dig(*([0] * 8999)).to_s).to eq '((abc))'
+    expect { @root.traverse(true) { |_, exp, _| exp } }.not_to raise_error
+    expect { @root.each_expression(true) { |exp| exp } }.not_to raise_error
+    expect { @root.each_expression(true) { |exp, _idx| exp } }.not_to raise_error
+  end
 end