+++ tasks/benchmarks/uri_regexp.rb
@@ -1,16 +1,12 @@
 # frozen_string_literal: true
 
-require 'benchmark/ips'
-require_relative '../../lib/regexp_parser'
+require_relative 'shared'
 
 puts 'Parsing a complex Regexp (URI.regexp)'
 
-require 'uri'
-regexp = URI::DEFAULT_PARSER.make_regexp
-
 Benchmark.ips do |x|
-  x.report('Scanner::scan') { Regexp::Scanner.scan(regexp) }
-  x.report('Lexer::lex')    { Regexp::Lexer.lex(regexp)    }
-  x.report('Parser::parse') { Regexp::Parser.parse(regexp) }
+  x.report('Scanner::scan') { Regexp::Scanner.scan(URI_REGEXP) }
+  x.report('Lexer::lex')    { Regexp::Lexer.lex(URI_REGEXP)    }
+  x.report('Parser::parse') { Regexp::Parser.parse(URI_REGEXP) }
   x.compare!
 end