+++ CHANGELOG.md
@@ -7,12 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- greatly improved scanning, lexing, and parsing performance
+  * ~5x for a single-expression patterns, ~2x for a 500-expression pattern
+- improved performance of `#each_expression` and `#traverse`
+- greatly improved performance when comparing parse results (via `#==`)
+- improved scanner error messages for invalid patterns, e.g. `'\xZ'`
+
 ### Fixed
 
+- `SystemStackError` when parsing or traversing very deeply nested trees (#109)
+- minor memory leak when accepting arbitrary user input as syntax version (#109)
 - `#match_length` now respects the quantifier of backrefs and subexp calls (#107)
-  * thanks to [Oskar Eichler](https://github.com/OskarEichler) for the report
 - `#match_length` now handles backref/subexp call multiplexing (#107)
-  * thanks to [Oskar Eichler](https://github.com/OskarEichler) for the report
+- thanks to [Oskar Eichler](https://github.com/OskarEichler) for reporting these
 
 ## [2.12.0] - 2026-04-04 - Janosch Müller